requirements.txt
psycopg2-binary
pandas
sqlalchemy
numpy
matplotlib
seaborn